+++ react/features/shared-video/functions.js
@@ -54,3 +54,15 @@ export function isVideoPlaying(stateful: Object | Function): boolean {
     return videoPlaying;
 }
 
+/**
+ * Fetch stopped video url
+ * @returns {void}
+ */
+export function fetchStoppedVideoUrl() {
+    
+    const state = APP.store.getState();
+    const sharedVideoCurrentState = state['features/shared-video'];
+
+    APP.API.notifySharedVideoStopped(sharedVideoCurrentState.videoUrl);
+
+}