+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.45] - 2023-04-12
+### Changed
+- Video owner update notification endpoint is extended
+- Endpoint for shared video stop notification is implemented
+
 ## [1.1.44] - 2023-02-23
 ### Changed
 - Minified external API js files updated in order to resolve the problem with the missing video sharing endpoints on the external api